agent1/requirements.txt
confluent_kafka
python-dotenv